requirements.txt
streamlit
kiteconnect
pandas
python-dotenv
pytz
watchdog
